402: chore: update mayastor pool CR definition file r=rohan2794 a=rohan2794

This PR update the mayastor pool CR definition file from control plane repo   https://github.com/mayadata-io/mayastor-control-plane/blob/develop/chart/templates/mayastorpoolcrd.yaml ) and fixes primitive msp state test due to CR changes.

JIRA: https://mayadata.atlassian.net/browse/MQ-2680

+++ src/primitive_msp_state/util.go
@@ -98,7 +98,7 @@ func verifyMspState(poolName string, crPool v1alpha1.MayastorPool,
 func verifyMspCapacity(poolName string, crPool v1alpha1.MayastorPool,
 	grpcPool mayastorclient.MayastorPool) bool {
 	var status bool
-	if crPool.Status.Capacity == int64(grpcPool.Capacity) {
+	if crPool.Status.Capacity == grpcPool.Capacity {
 		status = true
 	} else {
 		logf.Log.Info("verifyMspState",
@@ -114,7 +114,7 @@ func verifyMspCapacity(poolName string, crPool v1alpha1.MayastorPool,
 func verifyMspUsedSpace(poolName string, crPool v1alpha1.MayastorPool,
 	grpcPool mayastorclient.MayastorPool) bool {
 	var status bool
-	if crPool.Status.Used == int64(grpcPool.Used) {
+	if crPool.Status.Used == grpcPool.Used {
 		status = true
 	} else {
 		logf.Log.Info("verifyMspUsedSpace", "pool", poolName, "CR", crPool.Status.Used, "gRPC", grpcPool.Used)
@@ -166,7 +166,7 @@ func (c *mspStateConfig) checkPoolUsedSize(poolName string, replicaSize int64) e
 		time.Sleep(time.Duration(sleepTimeSecs) * time.Second)
 		pool, err := k8stest.GetMsPool(poolName)
 		Expect(err).ToNot(HaveOccurred(), fmt.Sprintf("failed to get mayastor pool %s %v", poolName, err))
-		if pool.Status.Used == replicaSize {
+		if int64(pool.Status.Used) == replicaSize {
 			return nil
 		}
 	}
